SquirrelPAK MC12 Release 1.5.3
Original Release Date - 6/27/2023

================================================================================
SquirrelPAK MC12 Change Log
================================================================================

SquirrelPAK MC12 Release 1.5.3

File Version 1.5.3+1.12.2

================================================================================

There was issues with the dev build profile and mods got added
that should not have gotten added and files were all wrong :-(

We took 1.4.9 fresh files and updated all the mods that had
updates and re-added things that were added to 1.5.0
So this should be a fresh version that should have
been 1.5.1.

##############################################################
!!WARNING - ATOMIC SCIENCE WILL BE REMOVED IN 1.6 - WARNING!!!
##############################################################

Also note we will be removing atomic science mod! This is due
to bugs found. We will remove it in the next update 1.6. If you
want to keep a world going you need to purge any Atomic Science
blocks and items from your world before upgrading to 1.6 to b
safe OR make a backup

We MAY replace Atomic Science with NuclearCraft: Overhauled

##############################################################
!!WARNING - ATOMIC SCIENCE WILL BE REMOVED IN 1.6 - WARNING!!!
##############################################################

Alert!!: Need at least 4GB of RAM!

Please report any issues with the mod pack on the issue tracker
on our github page

Changes from v1.4.9

Added:
------------------------------------------- 
starterkit_1.12.2-2.4
------------------------------------------- 

Updated:
------------------------------------------- 
AdvancedRocketry-1.12.2-1.7.0-232 to 1.12.2-2.0.0-13
AEAdditions-1.12.2-1.3.4 to 1.12.2-1.3.8
BiblioCraft[v2.4.5][MC1.12.2] to [v2.4.6][MC1.12.2]
Controlling-3.0.10 to 3.0.12.2
CraftTweaker2-1.12-4.1.20.673 to 2-1.12-4.1.20.689
CreativeCore_v1.10.69_mc1.12.2 to v1.10.70_mc1.12.2
engineersdecor-1.12.2-1.1.3 to 1.12.2-1.1.5
FastWorkbench-1.12.2-1.7.3 to 1.12.2-1.7.4
foamfix-0.10.14-1.12.2 to 0.10.15-1.12.2
rsgauges-1.12.2-1.2.5 to 1.12.2-1.2.8
MCTImmersiveTechnology-1.12.2-1.9.99 to 1.12.2-1.9.100
LibVulpes-1.12.2-0.4.2-87 to 1.12.2-0.4.2-25
LittleTiles_v1.5.57_mc1.12.2 to v1.5.72_mc1.12.2
modtweaker-4.0.20.5 to 4.0.20.11
OpenComputers-MC1.12.2-1.7.5.192 to 1.8.2+b4abbf9
OpenSecurity-1.12.2-1.0-83 to 1.12.2-1.0-93
reauth-3.6.0 to 4.0.7
SmoothFont-mc1.12.2-2.1.2 to 1.12.2-2.1.4
supermartijn642configlib-1.0.9-mc1.12 to 1.1.6-forge-mc1.12
supermartijn642corelib-1.0.16-mc1.12 to 1.1.10-forge-mc1.12
WanionLib-1.12.2-2.8 to 1.12.2-2.9
WarpDrive-1.12.2-1.5.23 to 1.12.2-1.5.28
wormhole-1.1.9-mc1.12 to 1.1.13
------------------------------------------- 

 
Removed:
------------------------------------------- 
None
-------------------------------------------